+++ CHANGES.md
@@ -6,7 +6,8 @@
 * [#33] Updated: `StreamEx.split("", pattern)` now returns stream of single `""` string instead of empty stream.
 * [#13] Added: `StreamEx.split` to split with single character delimiter.
 * [#35] Added: construction of all stream types from the `Iterator`.   
-* [#36] Added: `StreamEx.of(Enumeration)` static method.   
+* [#36] Added: `StreamEx.of(Enumeration)` static method.
+* [#32] Added: `MoreCollectors.collapsingNested` collector which collects elements to the list removing the nested elements.   
 * Updated documentation. 
 
 ### 0.5.0